watches/requirements.txt
amqp==5.2.0
APScheduler==3.6.3
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
cachetools==4.2.2
celery==5.4.0
certifi==2024.2.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2
django-filter==24.2
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
django-redis==5.4.0
gunicorn==22.0.0
kombu==5.3.7
packaging==24.0
phonenumbers==8.13.36
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg==3.1.8
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==13.7
pytz==2024.1
redis==5.0.4
six==1.16.0
sqlparse==0.5.0
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
vine==5.1.0
wcwidth==0.2.13
